version 1.1

+++ include/cesarCypher.h
@@ -8,10 +8,10 @@
 #define CESAR_CYPHER_H
 
 
-int cesarCypher(char *cypherText, char *plainText, int key);
-int rotaryCesarCypher(char *cypherText, char *plainText, int key);
-int cesarUnCypher(char *plainText, char *cypherText, int key);
-int rotaryCesarUnCypher(char *plainText, char *cypherText, int key);
+long int cesarCypher(char *cypherText, char *plainText,  long int key);
+long int rotaryCesarCypher(char *cypherText, char *plainText, long int key);
+long int cesarUnCypher(char *plainText, char *cypherText, long int key);
+long int rotaryCesarUnCypher(char *plainText, char *cypherText, long int key);
 
 #endif
